+++ tools/ice-scripts/bin/ice-build.js
@@ -4,7 +4,6 @@
 const program = require('commander');
 const cliInstance = require('../lib/utils/cliInstance');
 const build = require('../lib/build');
-const checkUpdater = require('../lib/utils/checkUpdater');
 
 /**
  * --project-type 参数说明
@@ -26,9 +25,7 @@ program
   )
   .parse(process.argv);
 
-checkUpdater().then(() => {
-  cliInstance.initByProgram(program);
-  build({
-    cliOptions: cliInstance.get(),
-  });
+cliInstance.initByProgram(program);
+build({
+  cliOptions: cliInstance.get(),
 });