+++ tools/ice-scripts/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.10.2
+
+- [chore] remove check-updater
+- [chore] add upgrade message
+- [chore] optimize collector
+
 ## 1.10.1
 
 - [fix] 组件&区块 build 指定 publicPath #1849